+++ packages/web-threejs/src/utils/Log.ts
@@ -1,5 +1,5 @@
 export function log(message?: any, ...optionalParams: any[]) {
-  if (process.env.REACT_APP_DEBUG && process.env.REACT_APP_DEBUG === 'true') {
+  if (false) {
     console.log(message, ...optionalParams);
   }
 }