+++ packages/web-threejs/CHANGELOG.md
@@ -1,8 +1,13 @@
-## [1.1.0] - 2024-02-01
+## [Unreleased]
 
 ### Changed
 
 - Updated RPM API to allow for current Ready Player Me characters
+
+## [1.1.0] - 2024-02-01
+
+### Changed
+
 - Updated README.md documentation for assets installation process
 - Updated LICENSE.md license information